src/main/kotlin/br/com/rhuan/util/ValidationUtil.kt
package br.com.rhuan.util

import br.com.rhuan.ProductServiceRequest

class ValidationUtil {
    companion object {
        fun validatePayload(payload: ProductServiceRequest?): ProductServiceRequest {
            payload?.let {
                if (it.name.isNullOrBlank())
                    throw IllegalArgumentException("Nome não pode ser nulo ou vazio")

                if (it.price.isNaN() || it.price < 0)
                    throw IllegalArgumentException("Preço precisa ser um valor valido")

                return it
            }
            throw IllegalArgumentException()
        }
    }
}